Add - add player header related to client, player and team structure Add - add position header related to vec2i and direction_t Fix - resource are uint16_t Add - get_client_cell function Add - get_player_list_at function Fix - fix forward position change Fix - fix eject position change Add - wrapped_coord function returning the correct position on the map

+++ sources/server/map/cell_management.c
@@ -46,7 +46,7 @@ bool remove_resource_on_cell(cell_t *cell, resource_t res, uint16_t nb)
 	return (true);
 }
 
-int get_resource_on_cell(cell_t const *const cell, resource_t res)
+uint16_t get_resource_on_cell(cell_t const *const cell, resource_t res)
 {
 	if (!cell) {
 		fprintf(stderr, "invalid pointer\n");
